+++ src/main/java/ojosama/talkak/member/controller/MemberController.java
@@ -5,7 +5,7 @@ import ojosama.talkak.member.dto.AdditionalInfoRequest;
 import ojosama.talkak.member.dto.AdditionalInfoResponse;
 import ojosama.talkak.member.dto.MyPageInfoRequest;
 import ojosama.talkak.member.dto.MyPageInfoResponse;
-import ojosama.talkak.member.service.MyPageService;
+import ojosama.talkak.member.service.MemberService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
@@ -19,21 +19,21 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api")
 @RequiredArgsConstructor
-public class MyPageController {
+public class MemberController {
 
-    private final MyPageService myPageService;
+    private final MemberService memberService;
 
     @GetMapping("/me")
     public ResponseEntity<MyPageInfoResponse> getMemberInfo() {
         Long memberId = 1L; //추후 로그인 기능 구현 시 수정 예정
-        MyPageInfoResponse memberInfo = myPageService.getMemberInfo(memberId);
+        MyPageInfoResponse memberInfo = memberService.getMemberInfo(memberId);
         return new ResponseEntity<>(memberInfo, HttpStatus.OK);
     }
 
     @PutMapping("/me")
     public ResponseEntity<MyPageInfoResponse> updateMemberInfo(@RequestBody MyPageInfoRequest myPageInfoRequest) {
         Long memberId = 1L;
-        MyPageInfoResponse memberInfo = myPageService.updateMemberInfo(memberId, myPageInfoRequest);
+        MyPageInfoResponse memberInfo = memberService.updateMemberInfo(memberId, myPageInfoRequest);
         return new ResponseEntity<>(memberInfo, HttpStatus.OK);
     }
 
@@ -41,7 +41,7 @@ public class MyPageController {
     public ResponseEntity<AdditionalInfoResponse> updateAdditionalInfo(@RequestBody AdditionalInfoRequest request, Authentication authentication) {
 //        Long id = Long.valueOf(authentication.getPrincipal().toString());
         Long id = 1L;
-        AdditionalInfoResponse response = myPageService.updateAdditionalInfo(id, request);
+        AdditionalInfoResponse response = memberService.updateAdditionalInfo(id, request);
         return ResponseEntity.ok().body(response);
     }
 }
